requirements.txt
# Core
tensorflow>=2.12,<3.0
numpy>=1.23
matplotlib>=3.7
pillow>=9.5
# App
streamlit>=1.31
# Optional, nice to have
rich>=13.7
